Safari-609.1.11.1/LayoutTests/fast/xpath/nsresolver-exception-expected.txt

CONSOLE MESSAGE: line 14: Error: this is an exception
SUCCESS
